requirements.txt
Flask>=2.2.3
Flask-CORS==3.0.10
gunicorn==20.1.0
Pillow==10.2.0
imageio[ffmpeg]==2.14.0
pillow-heif==0.5.5
pillow-avif==0.6.1